Add docker entrypoint script
Adds docker entrypoint script which creates opx-dev user and group
entries matching the uid and gid of the invoking user. The uid and
gid are passed from the host to the container through environment
variables LOCAL_UID and LOCAL_GID.

This provides a convienent way to have the container use the same
uid and gid for mapped filesystem access.