Work around overlayfs/pbuilder issue
Directory renames fail with a cross-device link error if the directory
is on a lower filesystem layer.  Userspace programs are to handle this
case, but as of this writing pbuilder does not and chroot updates fail.

As a work around, combine all steps of chroot creation in one "docker
run" invocation.