Fix builds in NFS mounts
Root is not allowed to write to NFS mounts. Pbuilder hooks run as root.

This change moves all commands that write to /mnt out of pbuilder hooks.
Pbuilder has a feature to run hooks as a user, but that user can not be
the same as the user calling pbuilder.

Change-Id: I886157e9e418c4cb4dbdc3aac9f21c9604b33dd1
Signed-off-by: Tyler Heucke <tyler@heucke.io>