Refactor Docker image
Instead of producing one Docker image per OPX release per Debian
release, we now produce one Docker image per Debian release. Only one
OPX release is preinstalled: unstable. If a different release is
specified, the pbuilder chroot is created on the fly. This method was
chosen for practicality. Creating a new chroot only takes 7.5 seconds,
but uses up 500MB. This method keeps the image slim, yet performant.

Additionally, the tag for the image is now tied to the commit of the
repository. The tag's format is now "${sha}-${debian_dist}"
(deadbeef-jessie, for examplt).

Change-Id: Ia0514655169574fff3262986f637dab0b76c0d77
Signed-off-by: Tyler Heucke <tyler.heucke@dell.com>